+++ apps/next-app/pages/modules/[questionBank]/tests/[testId]/exam.page.tsx
@@ -1,7 +1,20 @@
 import { useRouter } from "next/router";
-import { TestExam } from "@cf/next/tests";
+import {
+  DialogContent,
+  DialogTitle,
+  Drawer,
+  Modal,
+  ModalClose,
+  ModalDialog,
+} from "@mui/joy";
+import {
+  TestFinisher,
+  TestHeader,
+  TestNavigation,
+  TestQuestion,
+} from "@cf/next/tests";
 import { UserBugReport } from "@cf/next/user";
-import { AppHead } from "@cf/react/components";
+import { AppHead, useDisclose } from "@cf/react/components";
 import { ThemeOverrideColorScheme } from "@cf/react/theme";
 import { ssrHandler } from "@cf/trpc/server";
 import type { QuestionBankName } from "@cf/core/question-bank";
@@ -19,29 +32,63 @@ type Params = {
 
 export const Page: NextPage<Props> = ({ testId, questionBank }) => {
   const router = useRouter();
+  const questionNavSidebar = useDisclose();
+  const finishModal = useDisclose();
+
+  const navigateToReview = () => {
+    router.push(`/modules/${questionBank}/tests/${testId}/review`);
+  };
 
   return (
     <>
       <AppHead />
       <ThemeOverrideColorScheme questionBank={questionBank} />
-      <TestExam
+      <TestHeader
         noSsr
         testId={testId}
-        onTestFinished={() => {
-          router.push(`/modules/${questionBank}/tests/${testId}/review`);
-        }}
+        timerMode="timeSpent"
+        onQuestionsClicked={questionNavSidebar.open}
+        onFinishClicked={finishModal.open}
       />
+      <TestQuestion noSsr mode="exam" testId={testId} />
+
+      <Drawer
+        anchor="right"
+        open={questionNavSidebar.isOpen}
+        onClose={questionNavSidebar.close}
+        sx={{ "--Drawer-horizontalSize": "460px" }}
+      >
+        <ModalClose />
+        <DialogTitle>Question Navigation</DialogTitle>
+        <DialogContent sx={{ p: 2 }}>
+          <TestNavigation
+            noSsr
+            testId={testId}
+            onNavigation={questionNavSidebar.close}
+            onTestFinished={navigateToReview}
+          />
+        </DialogContent>
+      </Drawer>
+
+      <Modal open={finishModal.isOpen} onClose={finishModal.close}>
+        <ModalDialog>
+          <ModalClose />
+          <DialogContent sx={{ p: 2 }}>
+            <TestFinisher
+              noSsr
+              testId={testId}
+              onTestFinished={navigateToReview}
+            />
+          </DialogContent>
+        </ModalDialog>
+      </Modal>
       <UserBugReport />
     </>
   );
 };
 
 export const getServerSideProps = ssrHandler<Props, Params>(
-  async ({ helper, params }) => {
-    await TestExam.getData({ helper, params });
-    await UserBugReport.getData({ helper, params });
-    return { props: params };
-  },
+  async ({ params }) => ({ props: params }),
 );
 
 export default Page;
