+++ apps/next-app/pages/modules/[questionBank]/tests/[testId]/study.page.tsx
@@ -1,8 +1,21 @@
 import { useRouter } from "next/router";
-import { MissingPathParameter } from "@cf/base/errors";
-import { TestStudy } from "@cf/next/tests";
+import {
+  DialogContent,
+  DialogTitle,
+  Drawer,
+  Modal,
+  ModalClose,
+  ModalDialog,
+} from "@mui/joy";
+import {
+  TestFinisher,
+  TestHeader,
+  TestNavigation,
+  TestQuestion,
+  TestQuestionMeta,
+} from "@cf/next/tests";
 import { UserBugReport } from "@cf/next/user";
-import { AppHead } from "@cf/react/components";
+import { AppHead, useDisclose } from "@cf/react/components";
 import { ThemeOverrideColorScheme } from "@cf/react/theme";
 import { ssrHandler } from "@cf/trpc/server";
 import type { QuestionBankName } from "@cf/core/question-bank";
@@ -20,36 +33,80 @@ type Params = {
 
 export const Page: NextPage<Props> = ({ testId, questionBank }) => {
   const router = useRouter();
+  const metaSidebar = useDisclose();
+  const questionNavSidebar = useDisclose();
+  const finishModal = useDisclose();
+
+  const navigateToReview = () => {
+    router.push(`/modules/${questionBank}/tests/${testId}/review`);
+  };
+
+  const navigateToTestSearch = () => {
+    router.push(`/modules/${questionBank}/tests`);
+  };
 
   return (
     <>
       <AppHead />
       <ThemeOverrideColorScheme questionBank={questionBank} />
-      <TestStudy
+      <TestHeader
         noSsr
         testId={testId}
-        onTestFinished={() => {
-          router.push(`/modules/${questionBank}/tests/${testId}/review`);
-        }}
-        onTestDismissed={() => {
-          router.push(`/modules/${questionBank}/tests`);
-        }}
+        timerMode="timeSpent"
+        onQuestionsClicked={questionNavSidebar.open}
+        onMetaClicked={metaSidebar.open}
+        onFinishClicked={finishModal.open}
       />
+      <TestQuestion noSsr mode="study" testId={testId} />
+      <Drawer
+        anchor="right"
+        open={metaSidebar.isOpen}
+        onClose={metaSidebar.close}
+        sx={{ "--Drawer-horizontalSize": "640px" }}
+      >
+        <ModalClose />
+        <DialogTitle>Question Meta</DialogTitle>
+        <DialogContent sx={{ p: 2 }}>
+          <TestQuestionMeta noSsr testId={testId} questionBank={questionBank} />
+        </DialogContent>
+      </Drawer>
+      <Drawer
+        anchor="right"
+        open={questionNavSidebar.isOpen}
+        onClose={questionNavSidebar.close}
+        sx={{ "--Drawer-horizontalSize": "460px" }}
+      >
+        <ModalClose />
+        <DialogTitle>Question Navigation</DialogTitle>
+        <DialogContent sx={{ p: 2 }}>
+          <TestNavigation
+            noSsr
+            testId={testId}
+            onNavigation={questionNavSidebar.close}
+            onTestFinished={navigateToReview}
+          />
+        </DialogContent>
+      </Drawer>
+      <Modal open={finishModal.isOpen} onClose={finishModal.close}>
+        <ModalDialog>
+          <ModalClose />
+          <DialogContent sx={{ p: 2 }}>
+            <TestFinisher
+              noSsr
+              testId={testId}
+              onTestDismissed={navigateToTestSearch}
+              onTestFinished={navigateToReview}
+            />
+          </DialogContent>
+        </ModalDialog>
+      </Modal>
       <UserBugReport />
     </>
   );
 };
 
 export const getServerSideProps = ssrHandler<Props, Params>(
-  async ({ helper, params }) => {
-    const testId = params?.testId;
-    const questionBank = params?.questionBank;
-    if (!testId) throw new MissingPathParameter("testId");
-    if (!questionBank) throw new MissingPathParameter("questionBank");
-    await TestStudy.getData({ helper, params });
-    await UserBugReport.getData({ helper, params });
-    return { props: { testId, questionBank } };
-  },
+  async ({ params }) => ({ props: params }),
 );
 
 export default Page;
