--- libs/next/tests/src/containers/test-exam/index.ts
@@ -1,1 +1,0 @@
-export { TestExam } from "./test-exam";
