+++ libs/next/tests/src/containers/test-header/index.ts
@@ -1,0 +1,1 @@
+export { TestHeader } from "./test-header";
