+++ libs/next/tests/src/containers/test-navigation/index.ts
@@ -1,0 +1,1 @@
+export { TestNavigation } from "./test-navigation";
