+++ libs/next/tests/src/containers/test-question-meta/index.ts
@@ -1,0 +1,1 @@
+export { TestQuestionMeta } from "./test-question-meta";
