+++ libs/next/tests/src/containers/test-question/index.ts
@@ -1,0 +1,1 @@
+export { TestQuestion } from "./test-question";
