--- libs/next/tests/src/containers/test-study/index.ts
@@ -1,1 +1,0 @@
-export { TestStudy } from "./test-study";
