+++ libs/next/tests/src/index.ts
@@ -1,5 +1,8 @@
-export { TestExam } from "./containers/test-exam";
 export { TestMaker } from "./containers/test-maker";
 export { TestReview } from "./containers/test-review";
 export { TestSearch } from "./containers/test-search";
-export { TestStudy } from "./containers/test-study";
+export { TestHeader } from "./containers/test-header";
+export { TestQuestion } from "./containers/test-question";
+export { TestQuestionMeta } from "./containers/test-question-meta";
+export { TestNavigation } from "./containers/test-navigation";
+export { TestFinisher } from "./containers/test-finisher";
